+++ project/framework_test/src/rendering/vulkan/VulkanPipeline.h
@@ -15,7 +15,7 @@ public:
     vk::UniquePipeline pipeline;
 
     VulkanPipeline(vk::Device device, vk::RenderPass renderPass,
-                   Size<size_t> viewportSize,
+                   Size<uint32_t> viewportSize,
                    const VertexShader& vertex, const FragmentShader& fragment,
                    const vk::DescriptorSetLayout* descriptorSetLayout = nullptr, const vk::PushConstantRange* pushConstantRange = nullptr);
     VulkanPipeline(VulkanPipeline&&) noexcept = default;
